% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_integrated_graph.R
\name{adj_mat_checks}
\alias{adj_mat_checks}
\title{Perform checks on adjacency matrix inputs}
\usage{
adj_mat_checks(x)
}
\arguments{
\item{x}{adjacency matrix}
}
\value{
adjacency matrix
}
\description{
Perform checks on adjacency matrix input to ensure proper data format.
}
