% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_integrated_graph.R
\name{all.good}
\alias{all.good}
\title{Check if values of a vector are non-negative with at least one positive value.}
\usage{
\method{all}{good}(x)
}
\arguments{
\item{x}{numeric vector}
}
\value{
logical: if all values are non-negative with at least one positive value.
}
\description{
Check if values of a vector are non-negative with at least one positive value.
}
