% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RandomWalk.R
\name{column_normalize}
\alias{column_normalize}
\title{Column-normalize an adjacency matrix}
\usage{
column_normalize(adjM, norm, k, dim_names)
}
\arguments{
\item{adjM}{Adjacency matrix}

\item{norm}{Normalization scheme. One of 'degree' or 'modified_degree'}

\item{k}{Exponent applied to node degrees. Should be positive, since degrees are raised to -\emph{k}.}

\item{dim_names}{Dimnames of adjacency matrix}
}
\value{
Column-normalized matrix.
}
\description{
Given the normalization scheme, this computes a column-normalized adjacency matrix.

When norm='degree', the columns of the adjacency matrix are divided by their column sums. When norm='modified_degree', the adjacency matrix is first modified to decrease edge weights in proportion to the degree of its adjacent nodes, then it is column-normalized as with norm='degree'.
}
