% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{core_cc}
\alias{core_cc}
\title{Function for calculating the core-clustering coefficients of a graph}
\usage{
core_cc(g)
}
\arguments{
\item{g}{(igraph object): input graph. must be connected}
}
\value{
vector of core-clustering coefficients for each node
}
\description{
text
}
\details{
text
}
\examples{
core_clust_coefs = AMEND:::core_cc(glut4_graph)
head(core_clust_coefs)

}
