% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_AMEND.R
\name{create_aggregated_graph}
\alias{create_aggregated_graph}
\title{Create an aggregated graph}
\usage{
create_aggregated_graph(graph, control)
}
\arguments{
\item{graph}{igraph object. The multiplex graph to be aggregated.}

\item{control}{A named list. The element 'primary' contains the name of the primary layer for a multiplex component to be used during subnetwork identification. The element 'agg.method' contains a character scalar referring to an aggregation function.}
}
\value{
igraph object
}
\description{
An aggregated graph, in a multiplex context, is when the layers of a multiplex component are collapsed such that there is only one layer. A layer is designated as the primary, and edges in this layer take priority over edges of other layers. The collapsed multiplex will contain all of the original nodes.

The purpose is to aggregate a multiplex component before the subnetwork identification step (using \code{heinz()}) so that the algorithm considers the edges of the primary layer. This process involves aggregating node attribute values and fusing nodes together such that the fused node is adjacent to all of the edges of its predecessors.
}
\seealso{
\code{\link[=melt_graph]{melt_graph()}}, \code{\link[=expand_graph]{expand_graph()}}, \code{\link[=run_AMEND]{run_AMEND()}}
}
