% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_integrated_graph.R
\name{create_multiplex}
\alias{create_multiplex}
\title{Link common nodes between layers of multiplex.}
\usage{
create_multiplex(adj_matrix, multiplex.comps, data.type, brw.attr.nm)
}
\arguments{
\item{adj_matrix}{list of adjacency matrices.}

\item{multiplex.comps}{character string of names of unique components that are multiplex.}

\item{data.type}{character string of names of vertex attributes to preserve}

\item{brw.attr.nm}{character string  of vertex attribute name for biased random walk.}
}
\value{
A list of fused multiplex components
}
\description{
Given a list of inter-layer adjacency matrices (possibly from multiple multiplex components), this creates edges between common nodes across layers and places inter- and intra-layer matrices of a component in the same matrix.
}
