% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{data_preprocessing}
\alias{data_preprocessing}
\title{Pre-processing of experimental data for input into AMEND and RWR}
\usage{
data_preprocessing(
  graph,
  data.type = c("ECI", "logFC", "p_val", "binary", "other"),
  DOI = c("positive", "negative", "both"),
  seed.weight = 0.5,
  seed.scheme = c("zero_bottom", "zero_middle"),
  heterogeneous = FALSE
)
}
\arguments{
\item{graph}{Input graph}

\item{data.type}{One of ECI, logFC, p_val, or binary. If graph is specified, must have matching vertex attribute}

\item{DOI}{Direction of Interest for experimental values. One of negative, positive, or both. When "both" is specified, the absolute value is taken. Only relevant for data type ECI and logFC}

\item{seed.weight}{Relative weight to give to nodes not in the DOI in random walk with restart, between 0 and 1}

\item{seed.scheme}{Determines how the feature-wise experimental values are transformed for use as seed values in RWR. See Details.}

\item{heterogeneous}{Logical. If TRUE, network is considered heterogeneous (two distinct node types, e.g., proteins and metabolites). If TRUE, node_type must be included as an argument or graph vertex attribute}
}
\value{
igraph object with added 'seeds' and 'Z' vertex attributes
}
\description{
Pre-processing of experimental data for input into AMEND and RWR. For use inside \code{run_AMEND()}
}
\details{
Transforms experimental data for use with Random walk with restart (RWR), which requires non-negative seed values with at least one non-zero element. For subnetwork scoring, the experimental scores need to be standardized w.r.t. all nodes in the input network.
For data.type "ECI", the seeding scheme involves taking the absolute value of the ECIs then weighting those nodes NOT in the Direction of Interest (DOI) by 'seed.weight'
For data.type "logFC", the log fold changes are multiplied by -1 if the DOI is negative, 1 if the DOI is positive, and the absolute value is taken if the DOI is both. Then the values are exponentiated
For data.type "p_val", the p-values are transformed by -log10(p-value)
For data.type "binary", weight is assigned uniformly to "active" nodes, i.e., nodes with a value of 1 in the binary vector.
For data.type "other", no transformation is performed.
}
\examples{
# Attach igraph package
library(igraph)

# Inspect the igraph object included in AMEND package
# One can see it has vertex attributes name, symbol, and ECI
glut4_graph
plot(density(V(glut4_graph)$ECI))

# Transform experimental values (ECI) for multiple values of 'seed.weight'
new_graph1 = AMEND:::data_preprocessing(graph = glut4_graph, data.type = "ECI",
                                        DOI = "negative", seed.weight = 0.5)
new_graph2 = AMEND:::data_preprocessing(graph = glut4_graph, data.type = "ECI",
                                        DOI = "negative", seed.weight = 0.2)

# Comparing
plot(density(V(new_graph1)$seeds))
plot(density(V(new_graph2)$seeds))

plot(V(new_graph1)$ECI, V(new_graph1)$Z, type = "b",
     xlim = c(-1, 1), xlab = "ECI", ylab = "Standardized ECI")

}
