% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{dNetFind_alt}
\alias{dNetFind_alt}
\title{Heuristic solution of Maximum-weight Connected Subgraph problem}
\usage{
dNetFind_alt(g, scores)
}
\arguments{
\item{g}{(igraph object): input graph}

\item{scores}{(vector): named vector of scores for nodes in g}
}
\value{
subnetwork as an igraph object
}
\description{
Given a graph and a named vector of node scores, \code{dNetFind_alt()} heuristically finds a solution to the maximum-weight connected subgraph problem.
}
\details{
Details can be found in the \emph{dnet} package documentation \url{https://cran.r-project.org/web/packages/dnet/dnet.pdf}
}
\note{
Adapted from the \emph{dnet} package. Based on method from the \emph{BioNet} package
}
\examples{
library(igraph)

graph = sample_pa(n = 100, power = 1.2)
V(graph)$name = 1:100
g_scores = rnorm(100)
names(g_scores) = 1:100

new_g = AMEND:::dNetFind_alt(g = graph, scores = g_scores)

}
