% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{eci_scores}
\alias{eci_scores}
\title{ECI scores for proteins in the PPI network}
\format{
An object of class \code{numeric} of length 916.
}
\source{
\url{https://www.ncbi.nlm.nih.gov/geo/query/acc.cgi?acc=GSE35378}, \url{https://version-11-0b.string-db.org/cgi/download?sessionId=\%24input-\%3E\%7BsessionId\%7D&species_text=Mus+musculus}
}
\usage{
eci_scores
}
\description{
Equivalent Change Index (ECI) scores from a GLUT4 knockout-overexpression (KO-OX) microarray experiment in mice. The ECIs are ratio's of log fold changes from a differential expression analysis, in this case from the KO-Ctrl and OX-Ctrl comparisons
}
\keyword{datasets}
