% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{edgelists}
\alias{edgelists}
\title{List of edge lists representing a monoplex-heterogeneous network}
\format{
A named list with 3 elements:
\itemize{
\item 'prot': protein-protein interaction edge list
\item 'meta': chemical-chemical interaction edge list (meta=metabolite=chemical)
\item 'prot;meta': protein-chemical interaction edge list
}
}
\source{
Protein-protein interactions come from STRING. Chemical-chemical and chemical-protein interactions come from STITCH.
}
\usage{
edgelists
}
\description{
Protein-protein, chemical-chemical, and protein-chemical interactions, along with interaction confidence scores, in the form of edge lists. An edge list is a 2-column matrix with each row denoting an interaction. A possible third row can contain interaction scores.
}
\keyword{datasets}
