% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_integrated_graph.R
\name{el_checks}
\alias{el_checks}
\title{Perform checks on edge list inputs}
\usage{
el_checks(x)
}
\arguments{
\item{x}{edge list in form of matrix, data.frame, or data.table}
}
\value{
edge list
}
\description{
Perform checks on edge list input to ensure proper data format.
}
