% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RandomWalk.R
\name{entropy}
\alias{entropy}
\title{Compute entropy of a discrete probability distribution}
\usage{
entropy(x)
}
\arguments{
\item{x}{Probability vector}
}
\value{
Numeric value representing entropy
}
\description{
Computes entropy (with natural log) of a discrete probability distribution vector (i.e., sums to 1, values between 0 and 1).
}
