% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{exp.filtering.rate}
\alias{exp.filtering.rate}
\title{Exponential decay schedule for filtering rate}
\usage{
\method{exp}{filtering.rate}(eta0 = 0.5, d = 0.1)
}
\arguments{
\item{eta0}{(numeric): starting filtering rate}

\item{d}{(numeric): decay parameter}
}
\value{
vector of filtering rates for each iteration of AMEND
}
\description{
Exponential decay schedule for filtering rate. Used to shift the raw RWR scores
}
\examples{
rates = AMEND:::exp.filtering.rate(eta0 = 0.5, d = 0.2)
plot(1:100, rates)

}
