% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_AMEND.R
\name{exp_filtering_rate}
\alias{exp_filtering_rate}
\title{Exponential decay schedule for filtering rate}
\usage{
exp_filtering_rate(eta0 = 0.5, d = 0.1)
}
\arguments{
\item{eta0}{Starting filtering rate}

\item{d}{Decay parameter}
}
\value{
vector of filtering rates for each iteration of AMEND
}
\description{
Exponential decay schedule for filtering rate. Used to shift the raw RWR scores
}
\examples{
rates = AMEND:::exp_filtering_rate(eta0 = 0.5, d = 0.2)
plot(1:100, rates)

}
