% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_AMEND.R
\name{expand_graph}
\alias{expand_graph}
\title{Expand a graph}
\usage{
expand_graph(ig, ag, control)
}
\arguments{
\item{ig}{Input graph}

\item{ag}{Aggregated graph}

\item{control}{Named list. Information on aggregated graph.}
}
\value{
igraph object
}
\description{
Given an aggregated graph, returns its expanded counterpart, induced from the original, full graph.
Takes nodes from aggregated graph, identifies nodes that were in a collapsed multiplex component, duplicates them as necessary, and appends appropriate '|component_layer' information.
Returns an induced subgraph of ig that contains only node names matching nodes of aggregated graph.
}
