% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{find.decay}
\alias{find.decay}
\title{Sets the decay value}
\usage{
find.decay(N, eta0 = 0.5, n = 25)
}
\arguments{
\item{N}{(integer): Size of graph}

\item{eta0}{(numeric): starting filtering rate}

\item{n}{(integer): approximate size of final module}
}
\value{
decay value, numeric
}
\description{
Finds the max decay value for an exponential filtering rate schedule that will
allow the algorithm to arrive at a graph of size n.
}
\examples{
d = AMEND:::find.decay(N = 1000, eta0 = 0.5, n = 50)
d

}
