% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_AMEND.R
\name{find_decay}
\alias{find_decay}
\title{Sets the decay value}
\usage{
find_decay(N, eta0 = 0.5, n = 25, rate.diff = 0.07)
}
\arguments{
\item{N}{Size of graph}

\item{eta0}{Starting filtering rate}

\item{n}{Approximate size of final module}

\item{rate.diff}{Vector of differences between observed filtering rate and theoretical filtering rate}
}
\value{
decay value, numeric
}
\description{
Finds the max decay value for an exponential filtering rate schedule that will
allow the algorithm to arrive at a graph of size n.
}
\examples{
d = AMEND:::find_decay(N = 1000, eta0 = 0.5, n = 50)
d

}
