% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_AMEND.R
\name{get.eta0}
\alias{get.eta0}
\title{Set the starting filtering rate}
\usage{
get.eta0(g, n, method = c("difference", "ratio", "log.ratio"))
}
\arguments{
\item{g}{input graph}

\item{n}{Final module size to approximate}

\item{method}{One of 'difference', 'ratio', or 'log.ratio'}
}
\value{
Starting filtering rate
}
\description{
Sets the starting filtering rate (denoted as eta, the initial quantile of RWR scores used to shift scores before use in \code{heinz()}) in \code{run_AMEND()}.

eta is linearly increasing with D(N,n), where D is some distance metric between the input network size N and the final subnetwork size n.
For method = "difference", D(N,n) = N - n
For method = "ratio", D(N,n) = N / n
For method = "log.ratio", D(N,n) = log(N / n)
Below are the boundaries within which eta changes linearly with D(N,n). Set to capture "normal" or "expected" input network sizes and user-defined final module sizes.
N: (500, 15000), n: (5, 200), eta: (0.3, 0.8)
}
