% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get.type}
\alias{get.type}
\title{Get network location info from node labels}
\usage{
get.type(x, mode)
}
\arguments{
\item{x}{Node labels.}

\item{mode}{Integer. Type of node info to return. 1=node name. 2=component and layer names of nodes in \emph{x}. 3=component names of nodes in \emph{x}. 4=layer names of nodes in \emph{x}.}
}
\value{
Character vector of info specified by \emph{mode}, extracted from node labels.
}
\description{
Retrieves either the node name (\emph{mode}=1), the node component and layer (\emph{mode}=2), just the component (\emph{mode}=3), or just the layer (\emph{mode}=4).

Node labels follow the form 'name|component_layer'.
}
