% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_AMEND.R
\name{get_aggregate_method}
\alias{get_aggregate_method}
\title{Get a pre-defined aggregation function}
\usage{
get_aggregate_method(x)
}
\arguments{
\item{x}{character scalar}
}
\value{
aggregation function
}
\description{
Given a character string, this returns a function that will aggregate values of common nodes.
}
