% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_integrated_graph.R
\name{get_default_function}
\alias{get_default_function}
\title{Get a default function}
\usage{
get_default_function(f)
}
\arguments{
\item{f}{character string: 'binary', 'shift_scale', 'p_value', 'exp'. or NULL}
}
\value{
a function
}
\description{
Return a default function used to compute seed values for random walk with restart.

For the input \emph{f}, NULL means no transformation is done. 'binary' assumes that data is 0 or 1 so no transformation is done. 'shift_scale' is for data types with a range centered about zero and takes two parameters: \emph{DOI} (direction of interest: 1 for positive, -1 for negative, and 0 for both) and \emph{w} (numeric value between 0 and 1). It takes the absolute values of the data and then down-weights nodes that weren't in DOI by \emph{w} (if \emph{DOI}=0, \emph{w} is coerced to 1). 'p_value' assumes the data are p-values and calculates \emph{-log10()} . 'exp' assumes the data are log fold changes and has a \emph{DOI} arg. For \emph{DOI}=-1,1, it exponentiates the product of \emph{DOI} and \emph{logFC}. For \emph{DOI}=0, it exponentiates abs( \emph{logFC}).
}
