% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RandomWalk.R
\name{get_diagonal}
\alias{get_diagonal}
\title{Extract the diagonal elements of a matrix}
\usage{
get_diagonal(X)
}
\arguments{
\item{X}{matrix}
}
\value{
vector of diagonal elements of X.
}
\description{
Extract the diagonal elements of a square matrix.
}
