% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_integrated_graph.R
\name{get_layer}
\alias{get_layer}
\title{Get a layer of a mono/multiplex component.}
\usage{
get_layer(x, l)
}
\arguments{
\item{x}{matrix of a mono/multiplex component}

\item{l}{index of the layer to retrieve}
}
\value{
matrix
}
\description{
Get _l_th layer (in order of appearance) of a mono/multiplex component
}
