% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{glut4_adjM}
\alias{glut4_adjM}
\title{Mus musculus PPI network adjacency matrix}
\format{
An object of class \code{dgCMatrix} with 1033 rows and 1033 columns.
}
\source{
\url{https://www.ncbi.nlm.nih.gov/geo/query/acc.cgi?acc=GSE35378}, \url{https://version-11-0b.string-db.org/cgi/download?sessionId=\%24input-\%3E\%7BsessionId\%7D&species_text=Mus+musculus}
}
\usage{
glut4_adjM
}
\description{
Mus musculus PPI network adjacency matrix
}
\keyword{datasets}
