% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{glut4_graph}
\alias{glut4_graph}
\title{Mus musculus PPI network}
\format{
an igraph object containing an edge attribute "weight" and the following vertex attributes:
\itemize{
\item name: Ensmebl peptide ID
\item symbol: MGI symbols
\item ECI: Equivalent change indices (ECI)
}
}
\source{
\url{https://www.ncbi.nlm.nih.gov/geo/query/acc.cgi?acc=GSE35378}, \url{https://version-11-0b.string-db.org/cgi/download?sessionId=\%24input-\%3E\%7BsessionId\%7D&species_text=Mus+musculus}
}
\usage{
glut4_graph
}
\description{
A subset of the Mus musculus protein-protein interaction network, including both functional and physical interactions. Only interactions with a combined score greater than or equal to 0.8 were included
}
\keyword{datasets}
