% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_integrated_graph.R
\name{graph_checks}
\alias{graph_checks}
\title{Perform checks on igraph inputs}
\usage{
graph_checks(x)
}
\arguments{
\item{x}{igraph}
}
\value{
igraph
}
\description{
Perform checks on igraph input to ensure proper data format.
}
