% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/heinz.R
\name{heinz}
\alias{heinz}
\title{Heuristic solution of Maximum-weight Connected Subgraph problem}
\usage{
heinz(ig, scores, min.cluster.size = 2)
}
\arguments{
\item{ig}{Input graph}

\item{scores}{Named vector of scores for nodes in ig}

\item{min.cluster.size}{Minimum size of disconnected components of linker graph to consider when searching for maximum-weight subgraph. If NULL, only considers largest connected component of linker graph.}
}
\value{
a subnetwork as an igraph object
}
\description{
Given a graph and a named vector of node scores, \code{heinz()} heuristically finds a solution to the maximum-weight connected subgraph problem.
}
\details{
Details can be found in the \emph{dnet} package documentation under the \code{dnetfind()} function: \url{https://cran.r-project.org/web/packages/dnet/dnet.pdf}
}
\note{
Adapted from the \emph{dnet} package. Based on method from the \emph{BioNet} package
}
\examples{
library(igraph)

graph = sample_pa(n = 100, power = 1.2)
V(graph)$name = 1:100
g_scores = rnorm(100)
names(g_scores) = 1:100

new_g = heinz(ig = graph, scores = g_scores, min.cluster.size = 2)

}
\seealso{
\code{\link[=run_AMEND]{run_AMEND()}}, \code{\link[=RWR]{RWR()}}
}
