% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RandomWalk.R
\name{inflate_normalize}
\alias{inflate_normalize}
\title{Inflation-Normalization Procedure}
\usage{
inflate_normalize(nadjM)
}
\arguments{
\item{nadjM}{a column-normalized adjacency matrix (i.e., transition matrix)}
}
\value{
A modified transition matrix that has been adjusted for degree bias.
}
\description{
Directly modify the transition matrix to attenuate the influence of degree on diffusion scores (as evidenced by an increased entropy of stationary distribution associated with the modified transition matrix).
This is done by raising the values in each row of a left-stochastic transition matrix by an exponent greater than one which is a positive linear function of the stationary probability of the node associated with that row. This is followed by column normalization. This procedure displaces incoming transition probabilities to a node to the other outgoing transition probabilities of its neighbors as a function of degree.
}
