% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RandomWalk.R
\name{ipf}
\alias{ipf}
\title{Iterative Proportional Fitting}
\usage{
ipf(X, e = 1e-06)
}
\arguments{
\item{X}{Matrix}

\item{e}{Edge weight to add to zero-value diagonal elements of X to aid on convergence}
}
\value{
A named list
B: scaled matrix from the matrix product PXQ
p: diagonal elements of diagonal matrix P
q: diagonal elements of diagonal matrix Q

@export
}
\description{
Scales input matrix X to have row and column sums approximately equal to 1, thereby transforming X to a bistochastic (doubly stochastic) matrix.
}
