% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_integrated_graph.R
\name{largest_connected_component}
\alias{largest_connected_component}
\title{Get largest connected component of a graph}
\usage{
largest_connected_component(g)
}
\arguments{
\item{g}{igraph object}
}
\value{
igraph object of largest connected component.
}
\description{
Get largest connected component of a graph
}
