% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{list.of.adjmats}
\alias{list.of.adjmats}
\title{List of adjacency matrices representing a multiplex-homogeneous network}
\format{
A named list with 3 elements:
\itemize{
\item 'prot_1': adjacency matrix of layer 1 of protein multiplex component
\item 'prot_2': adjacency matrix of layer 2 of protein multiplex component
\item 'prot_3': adjacency matrix of layer 3 of protein multiplex component
}
}
\source{
Protein-protein interactions come from STRING.
}
\usage{
list.of.adjmats
}
\description{
3 adjacency matrices representing 3 layers of a multiplex protein-protein interaction network. The multiplex layers were taken from 3 different clusters of the full PPI network after running the Louvain clustering algorithm.
}
\keyword{datasets}
