% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{logFC_KO}
\alias{logFC_KO}
\title{Log fold changes for GLUT4 KO vs. Control for proteins in the PPI network}
\format{
An object of class \code{numeric} of length 1033.
}
\source{
\url{https://www.ncbi.nlm.nih.gov/geo/query/acc.cgi?acc=GSE35378}, \url{https://version-11-0b.string-db.org/cgi/download?sessionId=\%24input-\%3E\%7BsessionId\%7D&species_text=Mus+musculus}
}
\usage{
logFC_KO
}
\description{
Log fold changes for GLUT4 KO vs. Control for proteins in the PPI network
}
\keyword{datasets}
