% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RandomWalk.R
\name{modify_adj_mat}
\alias{modify_adj_mat}
\title{Scale edge weights of adjacency matrix by degrees of adjacent nodes.}
\usage{
modify_adj_mat(adjM, k)
}
\arguments{
\item{adjM}{Adjacency matrix}

\item{k}{Exponent applied to node degrees. Should be positive, since degrees are raised to -\emph{k}.}
}
\value{
Modified adjacency matrix
}
\description{
This functions creates a modified adjacency matrix. The adjacency matrix is left and right multiplied by two identical diagonal matrices whose diagonal elements are node degrees raised to the power -\emph{k}.

This decreases edge weights in proportion to the degrees of their adjacent nodes. In the context of column-normalization to obtain a transition matrix, this results in a biased random walk where the transition probability to a target node is inversely proportional to the target node's degree (all else being equal).

This can aid in mitigating degree bias associated with PPI networks.
}
