% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_integrated_graph.R
\name{n.char}
\alias{n.char}
\title{The number of a user-specified string in a character string.}
\usage{
n.char(x, ch)
}
\arguments{
\item{x}{character string or vector to search}

\item{ch}{character to search for}
}
\value{
integer
}
\description{
This returns the number of instances of \emph{ch} found in character string or vector \emph{x}.
}
