% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_integrated_graph.R
\name{n_layers}
\alias{n_layers}
\title{Get the number of layers for a mono/multiplex component}
\usage{
n_layers(x)
}
\arguments{
\item{x}{matrix}
}
\value{
integer
}
\description{
Get the number of layers for a mono/multiplex component
}
