% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{norm.adj}
\alias{norm.adj}
\title{Function for normalizing the adjacency matrix of a graph for input into random walk with restart (RWR)}
\usage{
norm.adj(g, adjM, norm = c("core", "degree", "weight"))
}
\arguments{
\item{g}{(igraph object): input graph}

\item{adjM}{(matrix): adjacency matrix}

\item{norm}{(character): normalization method}
}
\value{
normalized adjacency matrix
}
\description{
text
}
\details{
text
}
\examples{
library(igraph)

adjm = matrix(c(0, 1, 0, 0, 0, 0, 0, 0,
0, 0, 1, 1, 1, 0, 0, 0,
0, 0, 0, 0, 1, 1, 0, 0,
0, 0, 0, 0, 0, 0, 1, 0,
0, 0, 0, 0, 0, 0, 1, 1,
0, 0, 0, 0, 0, 0, 0, 0,
0, 0, 0, 0, 0, 0, 0, 0,
0, 0, 0, 0, 0, 0, 0, 0), nrow = 8)
g = graph_from_adjacency_matrix(adjm, mode = 'undirected')

adj_norm = AMEND:::norm.adj(g, adjm)

}
