% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RandomWalk.R
\name{stationary.distr}
\alias{stationary.distr}
\title{Compute stationary distribution of a transition matrix}
\usage{
stationary.distr(x)
}
\arguments{
\item{x}{Left-stochastic transition matrix}
}
\value{
Probability vector of the stationary distribution.
}
\description{
Compute the stationary distribution of a Markov chain represented by a transition matrix. This is the eigenvector associated with the absolute largest eigenvalue of the transition matrix (eigenvector normalized to sum to one).
}
