% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RandomWalk.R
\name{sum2one}
\alias{sum2one}
\title{Coerce columns of a matrix to sum to one}
\usage{
sum2one(X)
}
\arguments{
\item{X}{matrix}
}
\value{
A matrix, with all columns summing to one.
}
\description{
Divides columns by column sums so that all columns sum to one.
}
