% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_integrated_graph.R
\name{transform_bipartite}
\alias{transform_bipartite}
\title{Transform a bipartite graph to conform column-wise to a matrix and row-wise to another matrix.}
\usage{
transform_bipartite(col.mat, row.mat, bp.mat)
}
\arguments{
\item{col.mat}{matrix to which \emph{bp.mat} is matched column-wise}

\item{row.mat}{matrix to which \emph{bp.mat} is matched row-wise}

\item{bp.mat}{bipartite matrix}
}
\value{
a matrix
}
\description{
Transform bipartite graph \emph{bp.mat} to conform column-wise to \emph{col.mat} and row-wise to \emph{row.mat}
}
