% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{transition_matrix}
\alias{transition_matrix}
\title{Create a transition matrix from the adjacency matrix of a graph}
\usage{
transition_matrix(
  g,
  adjM,
  norm = c("degree", "modified_degree"),
  heterogeneous = FALSE,
  node_type = NULL,
  jump = 0.5,
  k = 0.5
)
}
\arguments{
\item{g}{Input graph}

\item{adjM}{Adjacency matrix}

\item{norm}{Normalization method}

\item{heterogeneous}{Logical. If TRUE, network is considered heterogeneous (two distinct node types, e.g., proteins and metabolites). If TRUE, node_type must be included as an argument or graph vertex attribute.}

\item{node_type}{Named vector of node types. Only considered if heterogeneous=TRUE. Can only accommodate two unique node types.}

\item{jump}{Probability of a random walker jumping from one network to the other in RWR (when heterogeneous=TRUE).}

\item{k}{Value between 0 and 1. When norm = "modified_degree", the adjacency matrix is first left and right multiplied by a diagonal matrix of node degrees, which is raised to the power -k. As k increases, edge weights are penalized more for the degrees of their adjacent nodes.}
}
\value{
transition matrix
}
\description{
\code{transition_matrix()} creates a transition matrix from the adjacency matrix of a graph for input into random walk with restart (RWR).
}
\details{
There are two normalization schemes available.
\itemize{
\item degree: Divides each column of the adjacency matrix by the column total
\item modified_degree: First creates a modified adjacency matrix by left and right multiplying the adjacency matrix by a diagonal matrix raised to the power of -k, whose diagonal elements are the column sums of the original adjacency matrix. Then proceed to column-normalize this modified adjacency matrix.
}
}
\examples{

library(igraph)

# Creating an adjacency matrix and graph
adjm = matrix(c(0, 1, 0, 0, 0, 0, 0, 0,
               1, 0, 1, 1, 1, 0, 0, 0,
               0, 1, 0, 0, 1, 1, 0, 0,
               0, 1, 0, 0, 0, 0, 1, 0,
               0, 1, 1, 0, 0, 0, 1, 1,
               0, 0, 1, 0, 0, 0, 0, 0,
               0, 0, 0, 1, 1, 0, 0, 0,
               0, 0, 0, 0, 1, 0, 0, 0), nrow = 8)
g = graph_from_adjacency_matrix(adjm, mode = 'undirected')

adj_norm = AMEND:::transition_matrix(g, adjm, norm = "degree")

}
