% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{core_cc}
\alias{core_cc}
\title{Calculate the core-clustering coefficients of nodes in a graph}
\usage{
core_cc(g, weight = TRUE)
}
\arguments{
\item{g}{Input graph. Must be connected}

\item{weight}{Logical. If true, the edge density is calculated taking into account edge weights}
}
\value{
Vector of core-clustering coefficients for each node
}
\description{
\code{core_cc()} calculates the core-clustering coefficients of all nodes in a graph. This network concept was introduced by Bader and Hogue (2003)
}
\details{
The core-clustering coefficient of a node is the density of the maximum k-core of the immediate neighborhood of that node. The k-core of a graph is the maximal subgraph in which every node has degree >= k.
}
\examples{
core_clust_coefs = AMEND:::core_cc(glut4_graph)
head(core_clust_coefs)

}
