% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{dNetFind_alt}
\alias{dNetFind_alt}
\title{Heuristic solution of Maximum-weight Connected Subgraph problem}
\usage{
dNetFind_alt(g, scores)
}
\arguments{
\item{g}{(igraph object): input graph}

\item{scores}{(vector): named vector of scores for nodes in g}
}
\value{
subnetwork as an igraph object
}
\description{
text
}
\details{
text
}
\note{
Adapted from the dnet package. Based on method from the BioNet package
}
\examples{
library(igraph)

graph = sample_pa(n = 100, power = 1.2)
V(graph)$name = 1:100
g_scores = rnorm(100)
names(g_scores) = 1:100

new_g = AMEND:::dNetFind_alt(g = graph, scores = g_scores)

}
