% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{dRWR_alt}
\alias{dRWR_alt}
\title{Random walk with restart (RWR) procedure.}
\usage{
dRWR_alt(g, nadjM, setSeeds = NULL, restart = 0.75)
}
\arguments{
\item{g}{(igraph object): input graph}

\item{nadjM}{(matrix): normalized adjacency matrix}

\item{setSeeds}{(vector): vector of seed values}

\item{restart}{(numeric): restart parameter}
}
\value{
vector of propagation scores from RWR
}
\description{
text
}
\details{
text
}
\note{
Adapted from the dnet package
}
\examples{
library(igraph)

adjm = matrix(c(0, 1, 0, 0, 0, 0, 0, 0,
0, 0, 1, 1, 1, 0, 0, 0,
0, 0, 0, 0, 1, 1, 0, 0,
0, 0, 0, 0, 0, 0, 1, 0,
0, 0, 0, 0, 0, 0, 1, 1,
0, 0, 0, 0, 0, 0, 0, 0,
0, 0, 0, 0, 0, 0, 0, 0,
0, 0, 0, 0, 0, 0, 0, 0), nrow = 8)
g = graph_from_adjacency_matrix(adjm, mode = 'undirected')
V(g)$name = 1:8

m = adjm \%*\% diag(degree(g))
adj_norm =  m \%*\% diag(1 / ifelse(colSums(m) == 0, 1, colSums(m)))
seeds = runif(8)
names(seeds) = 1:8

rwr = AMEND:::dRWR_alt(g, adj_norm, seeds, 0.8)

}
