% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{extract_string}
\alias{extract_string}
\title{Extract substrings before or after a delimiter}
\usage{
extract_string(x, k, pos)
}
\arguments{
\item{x}{vector of character strings}

\item{k}{Pattern to split on}

\item{pos}{position of the vector to return after split}
}
\value{
character string
}
\description{
This facilitates identifying the component and the layer where each node is located by splitting on '|' and '\verb{_}' delimiters. This component/layer information is appended to node names in the form '|component_layer'.
}
