% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_AMEND.R
\name{get_subnetwork}
\alias{get_subnetwork}
\title{Get a larger subnetwork from previous iterations of AMEND}
\usage{
get_subnetwork(ig, amend_object, k)
}
\arguments{
\item{ig}{Input graph}

\item{amend_object}{Result from \code{run_AMEND()}}

\item{k}{Iteration associated with the subnetwork you want to retrieve}
}
\value{
igraph object of the subnetwork associated with the given iteration.
}
\description{
This function allows the user to examine intermediate subnetworks that were generated during 'run_AMEND()', since the algorithm may overshoot the approximate final module size, or the user may be interested in the parent subnetworks of the final module.
}
\examples{
# Attach igraph library
library(igraph)

# Inspect the igraph object included in AMEND package.
# One can see it has vertex attributes name, symbol, and ECI
glut4_graph
head(V(glut4_graph)$ECI)

# Use run_AMEND() with an igraph object with a vertex attribute matching data.type arg
subnet1 = run_AMEND(graph = glut4_graph, n = 25, data = "ECI", FUN = "shift_scale",
                    FUN.params = list(DOI = -1, w = 0.5), normalize = "degree")

# Inspect the statistics of the intermediate subnetworks
subnet1$stats

# Retrieve the subnetwork of iteration 7
subnet7 = get_subnetwork(ig = glut4_graph, amend_object = subnet1, k = 7)
subnet7

}
