% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_integrated_graph.R
\name{melt_graph}
\alias{melt_graph}
\title{Melt a graph}
\usage{
melt_graph(g)
}
\arguments{
\item{g}{igraph}
}
\value{
igraph with vertex attributes preserved.
}
\description{
Melting a graph is finding nodes with the same node name and fusing them together so that there is just one node that is adjacent to all of the edges.
}
