% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_AMEND.R
\name{restart_grid_search}
\alias{restart_grid_search}
\title{Grid search for the restart parameter in RWR}
\usage{
restart_grid_search(
  ig,
  ag,
  n.adj.M,
  seeds,
  filtering_rate,
  heterogeneous = FALSE,
  multiplex = FALSE,
  net.weight,
  layer.weight,
  iteration = 1,
  agg.method = NULL,
  degree.bias = NULL,
  in.parallel = FALSE,
  n.cores
)
}
\arguments{
\item{ig}{Input graph}

\item{ag}{Aggregated graph}

\item{n.adj.M}{Normalized adjacency matrix}

\item{seeds}{A named vector of seed values}

\item{filtering_rate}{Quantile for shifting the raw RWR scores}

\item{heterogeneous}{Logical. If TRUE, network is considered heterogeneous (two distinct node types, e.g., proteins and metabolites). If TRUE, \emph{node_type} must be included as an argument or graph vertex attribute}

\item{multiplex}{Logical. If true, graph is assumed to contain multiplex components.}

\item{net.weight}{A named vector, or NULL. Relative weight given to nodes of a component of graph, applied to seed vector in RWR. Only used when heterogeneous=TRUE.}

\item{layer.weight}{A named list of named vectors, or NULL. Relative weight given to nodes of a layer of a component of graph, applied to seed vector in RWR. List element names correspond to multiplex components, and vector names correspond to layers within a multiplex.}

\item{iteration}{Current iteration of the AMEND algorithm.}

\item{agg.method}{A named list. The element 'primary' contains the name of the primary layer for a multiplex component to be used during subnetwork identification. The element 'agg.method' contains a character scalar referring to an aggregation function.}

\item{degree.bias}{List or NULL. NULL (default) for no degree bias adjustment. List names should include 'component', which is a character string of length greater than or equal to 1 specifying the components or layers to which the degree bias adjustment will be applied, and 'method', which is a character string of 'BS' (bistochastic scaling), 'IN' (inflation-normalization), or 'SDS' (stationary distribution scaling).}

\item{in.parallel}{Logical. Run computations in parallel.}

\item{n.cores}{number of cores to use in parallel computations}
}
\value{
a list containing an igraph object, subnetwork score, and restart value
}
\description{
Grid search for the restart parameter in RWR. For use inside \code{run_AMEND()}
}
\details{
This is a simple grid search for the restart probability parameter in RWR. For a grid of restart probability vaules, \code{RWR()} is run, raw scores are shifted by the current filtering rate of that iteration, then \code{heinz()} is run to find optimal subnetworks, which are scored. The restart probability resulting in largest subnetwork score is used.
}
\examples{
library(igraph)

# Creating an adjacency matrix and graph
adjm = matrix(c(0, 1, 0, 0, 0, 0, 0, 0,
               1, 0, 1, 1, 1, 0, 0, 0,
               0, 1, 0, 0, 1, 1, 0, 0,
               0, 1, 0, 0, 0, 0, 1, 0,
               0, 1, 1, 0, 0, 0, 1, 1,
               0, 0, 1, 0, 0, 0, 0, 0,
               0, 0, 0, 1, 1, 0, 0, 0,
               0, 0, 0, 0, 1, 0, 0, 0), nrow = 8, dimnames = list(1:8, 1:8))
g = graph_from_adjacency_matrix(adjm, mode = 'undirected')
V(g)$name = 1:8
plot(g)

# Normalizing the adjacency matrix
adj_norm = adjm \%*\% diag(1 / degree(g))
dimnames(adj_norm) = dimnames(adjm)

# Creating a named vector of seed values
seeds = runif(8)
names(seeds) = 1:8

search = AMEND:::restart_grid_search(ig = g, ag = g, n.adj.M = adj_norm,
                                     seeds = seeds, filtering_rate = 0.3)
plot(search[[1]])

}
