% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{RWR}
\alias{RWR}
\title{Random walk with restart (RWR) procedure}
\usage{
RWR(
  nadjM,
  setSeeds = NULL,
  restart = 0.75,
  heterogeneous = FALSE,
  node_type = NULL,
  net1.weight = 0.5
)
}
\arguments{
\item{nadjM}{Normalized adjacency matrix}

\item{setSeeds}{Vector of seed values}

\item{restart}{Restart parameter}

\item{heterogeneous}{Logical. If TRUE, network is considered heterogeneous (two distinct node types, e.g., proteins and metabolites). If TRUE, node_type must be included as an argument or graph vertex attribute}

\item{node_type}{Named vector of node types. Only considered if heterogeneous=TRUE. Can only accommodate two unique node types.}

\item{net1.weight}{Relative weight to give to nodes of network 1 (the first type to appear in node_type).}
}
\value{
vector of propagation scores from RWR
}
\description{
\code{RWR()} implements the RWR procedure. This simulates random walkers on a graph, with a certain probability of returning to the seeds nodes.
}
\examples{
library(igraph)

# Creating an adjacency matrix and graph
adjm = matrix(c(0, 1, 0, 0, 0, 0, 0, 0,
               1, 0, 1, 1, 1, 0, 0, 0,
               0, 1, 0, 0, 1, 1, 0, 0,
               0, 1, 0, 0, 0, 0, 1, 0,
               0, 1, 1, 0, 0, 0, 1, 1,
               0, 0, 1, 0, 0, 0, 0, 0,
               0, 0, 0, 1, 1, 0, 0, 0,
               0, 0, 0, 0, 1, 0, 0, 0), nrow = 8, dimnames = list(1:8, 1:8))
g = graph_from_adjacency_matrix(adjm, mode = 'undirected')
V(g)$name = 1:8

# Normalizing the adjacency matrix
adj_norm = adjm \%*\% diag(1 / degree(g))

# Creating a named vector of seed values
seeds = runif(8)
names(seeds) = 1:8

rwr = AMEND:::RWR(nadjM = adj_norm, setSeeds = seeds, restart = 0.8)

}
