% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{transition_matrix}
\alias{transition_matrix}
\title{Create a transition matrix from the adjacency matrix of a graph}
\usage{
transition_matrix(
  g,
  adjM,
  norm = c("degree", "core"),
  heterogeneous = FALSE,
  node_type = NULL,
  jump = 0.5
)
}
\arguments{
\item{g}{Input graph}

\item{adjM}{Adjacency matrix}

\item{norm}{Normalization method}

\item{heterogeneous}{Logical. If TRUE, network is considered heterogeneous (two distinct node types, e.g., proteins and metabolites). If TRUE, node_type must be included as an argument or graph vertex attribute}

\item{node_type}{Named vector of node types. Only considered if heterogeneous=TRUE. Can only accommodate two unique node types.}

\item{jump}{Probability of a random walker jumping from one network to the other in RWR (when heterogeneous=TRUE)}
}
\value{
transition matrix
}
\description{
\code{transition_matrix()} creates a transition matrix from the adjacency matrix of a graph for input into random walk with restart (RWR).
}
\details{
There are two normalization schemes available.
\itemize{
\item degree: divides each column of the adjacency matrix by the column total
\item core: uses the concept of node coreness. Details can be found in \url{https://academic.oup.com/nar/article/48/17/e98/5879427}
}
}
\examples{


library(igraph)

# Creating an adjacency matrix and graph
adjm = matrix(c(0, 1, 0, 0, 0, 0, 0, 0,
               1, 0, 1, 1, 1, 0, 0, 0,
               0, 1, 0, 0, 1, 1, 0, 0,
               0, 1, 0, 0, 0, 0, 1, 0,
               0, 1, 1, 0, 0, 0, 1, 1,
               0, 0, 1, 0, 0, 0, 0, 0,
               0, 0, 0, 1, 1, 0, 0, 0,
               0, 0, 0, 0, 1, 0, 0, 0), nrow = 8)
g = graph_from_adjacency_matrix(adjm, mode = 'undirected')

adj_norm = AMEND:::transition_matrix(g, adjm, norm = "degree")

}
\keyword{internal}
